FIX comments, rotation, direction
Fixed rotation and direction bugs and issues. Commented all scripts
