+++ unityfiles/Assets/Scripts/Game/Enemy/Behaviour/EnemyIdleBehaviour.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 440b31199088240c7a718cda73d542c3
+guid: 6c19fd53f39fe4173a7c05d91f3c847a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
